+++ src/propietarios/propietarios.service.ts
@@ -1,20 +1,18 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from '../prisma/prisma.service';
-import { CreatePropietarioDto } from './dto/create.propietario.dto';
-import { UpdatePropietarioDto } from './dto/update.propietario.dto';
+import { CreatePropietarioDto } from './dto/create-propietario.dto';
+import { UpdatePropietarioDto } from './dto/update-propietario.dto';
 
 @Injectable()
 export class PropietariosService {
   constructor(private prisma: PrismaService) {}
 
-  // CREATE
   create(createPropietarioDto: CreatePropietarioDto) {
     return this.prisma.propietarios.create({
       data: createPropietarioDto,
     });
   }
 
-  // READ ALL
   findAll() {
     return this.prisma.propietarios.findMany({
       include: {
@@ -23,34 +21,45 @@ export class PropietariosService {
     });
   }
 
-  // READ ONE
-  findOne(id: number) {
-    return this.prisma.propietarios.findUnique({
-      where: {
-        id_propietario: id,
-      },
-      include: {
-        mascotas: true,
-      },
+  async findOne(id: number) {
+    const propietario = await this.prisma.propietarios.findUnique({
+      where: { id_propietario: id },
+      include: { mascotas: true },
     });
+
+    if (!propietario) {
+      throw new NotFoundException('Propietario no encontrado');
+    }
+
+    return propietario;
   }
 
-  // UPDATE
-  updatePropietario(id: number, data: UpdatePropietarioDto) {
+  async updatePropietario(id: number, data: UpdatePropietarioDto) {
+    const propietario = await this.prisma.propietarios.findUnique({
+      where: { id_propietario: id },
+    });
+
+    if (!propietario) {
+      throw new NotFoundException('Propietario no existe');
+    }
+
     return this.prisma.propietarios.update({
-      where: {
-        id_propietario: id,
-      },
+      where: { id_propietario: id },
       data,
     });
   }
 
-  // DELETE
-  deletePropietario(id: number) {
+  async deletePropietario(id: number) {
+    const propietario = await this.prisma.propietarios.findUnique({
+      where: { id_propietario: id },
+    });
+
+    if (!propietario) {
+      throw new NotFoundException('Propietario no existe');
+    }
+
     return this.prisma.propietarios.delete({
-      where: {
-        id_propietario: id,
-      },
+      where: { id_propietario: id },
     });
   }
 }
